Server service GetAllAccessible returns server fragments
Server fragments are incomplete server objects which only contain the ID and Name and do not include any other sensitive info.

This is used in the GetAllAccessible server service method and the GetServers server handler method.

Server fragments are returned by the service when the user does NOT have permission to view the server but they do have permission to view either player records or infraction records of the server.